+++ openyich-framework/src/main/java/com/openyich/framework/boot/autoconfigure/OpenYichAutoConfiguration.java
@@ -13,6 +13,7 @@ import com.openyich.framework.boot.config.LocaleConfiguration;
 import com.openyich.framework.boot.config.SwaggerConfiguration;
 import com.openyich.framework.boot.config.WebConfigurer;
 import com.openyich.framework.boot.errors.ExceptionTranslator;
+import com.openyich.framework.boot.errors.ProblemResponseBodyAdvice;
 import com.openyich.framework.boot.hazelcast.HazelcastCacheConfiguration;
 import com.openyich.framework.boot.security.ssl.UndertowSSLConfiguration;
 import com.openyich.framework.boot.security.uaa.UaaConfiguration;
@@ -31,6 +32,7 @@ import com.openyich.framework.boot.security.uaa.UaaConfiguration;
   HttpHeadersTranslator.class,
   JacksonConfiguration.class,
   LocaleConfiguration.class,
+  ProblemResponseBodyAdvice.class,
   SwaggerConfiguration.class,
   UaaConfiguration.class,
   UndertowSSLConfiguration.class,