+++ openyich-framework/src/main/java/com/openyich/framework/boot/result/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Spring MVC Response Wrapper.
+ */
+package com.openyich.framework.boot.result;